+++ tests/testthat/tests.brm.R
@@ -6,7 +6,7 @@ test_that("brm produces expected errors", {
   dat <- data.frame(y = rnorm(10), x = rnorm(10), g = rep(1:5, 2))
   
   # formula parsing
-  expect_error(brm(~ x + (1|g), dat), 
+  expect_error(brm(~ x + (1|g), dat, file = "test"), 
                "response variable is missing")
   expect_error(brm(bf(y ~ a, nl = TRUE)),
                "No non-linear parameters specified")